Changes to First Level
Changes made according to feedback focusing on being more forgiving and less harsher punishments as this is the first level and will set a baseline of how difficult it could become for future levels.

Spaced out the platforms to prevent players from feeling cramping and being restricted in the movement.

Removed the last section that contain multiple small jumps as that was meant for a toucher challenge after reaching the 'end' and will now be a proper level.